requirements.txt
pyautogui
python-chess
berserk
